Refactor WinterStorageApplicationList to include its table
